Remove plumbing package
The plumbing package's scope has always been ill-defined. The functions
remaining had a single caller or were clearer to inline into the
surrounding code.
