+++ internal/buildpack/android_sdk.go
@@ -7,7 +7,6 @@ import (
 
 	"github.com/yourbase/yb"
 	"github.com/yourbase/yb/internal/biome"
-	"github.com/yourbase/yb/internal/plumbing"
 	"zombiezen.com/go/log"
 )
 
@@ -53,7 +52,7 @@ func installAndroidSDK(ctx context.Context, sys Sys, spec yb.BuildpackSpec) (bio
 	if data.OS == "" {
 		return biome.Environment{}, fmt.Errorf("unsupported os %s", desc.OS)
 	}
-	downloadURL, err := plumbing.TemplateToString(template, data)
+	downloadURL, err := templateToString(template, data)
 	if err != nil {
 		return biome.Environment{}, err
 	}
