+++ internal/buildpack/glide.go
@@ -5,7 +5,6 @@ import (
 
 	"github.com/yourbase/yb"
 	"github.com/yourbase/yb/internal/biome"
-	"github.com/yourbase/yb/internal/plumbing"
 	"zombiezen.com/go/log"
 )
 
@@ -35,7 +34,7 @@ func installGlide(ctx context.Context, sys Sys, spec yb.BuildpackSpec) (biome.En
 		OS:      desc.OS,
 		Arch:    desc.Arch,
 	}
-	downloadURL, err := plumbing.TemplateToString(template, data)
+	downloadURL, err := templateToString(template, data)
 	if err != nil {
 		return biome.Environment{}, err
 	}
