+++ internal/buildpack/protoc.go
@@ -6,7 +6,6 @@ import (
 
 	"github.com/yourbase/yb"
 	"github.com/yourbase/yb/internal/biome"
-	"github.com/yourbase/yb/internal/plumbing"
 	"zombiezen.com/go/log"
 )
 
@@ -46,7 +45,7 @@ func installProtoc(ctx context.Context, sys Sys, spec yb.BuildpackSpec) (biome.E
 	default:
 		return biome.Environment{}, fmt.Errorf("unsupported os/arch %s/%s", desc.OS, desc.Arch)
 	}
-	downloadURL, err := plumbing.TemplateToString(template, data)
+	downloadURL, err := templateToString(template, data)
 	if err != nil {
 		return biome.Environment{}, err
 	}
